refact: move application aggregate to application microservice
