+++ deploy/helm/applications/values.yaml
@@ -14,6 +14,7 @@ service:
 
 env:
   ASPNETCORE_URLS: "http://0.0.0.0:8080"
+  APPLICATIONS_MONGO_DB: "hireflow_applications"
 
 secretRefName: "hireflow-connections"
