+++ services/applications/Contracts/ScheduleInterviewRequest.cs
@@ -1,0 +1,8 @@
+namespace WastingNoTime.HireFlow.Applications.Contracts;
+
+// Request used to schedule an interview
+public sealed record ScheduleInterviewRequest(
+    DateTime ScheduledAtUtc,
+    int DurationMinutes,
+    string? Location
+);
