+++ services/applications/Messaging/INotificationsCommandBus.cs
@@ -1,0 +1,13 @@
+namespace WastingNoTime.HireFlow.Applications.Messaging;
+
+public interface INotificationsCommandBus : IAsyncDisposable
+{
+    Task PublishSendEmailAsync(
+        string to,
+        string subject,
+        string body,
+        string applicationId,
+        string interviewId,
+        long jobId,
+        CancellationToken ct = default);
+}
