+++ services/gateway/appsettings.json
@@ -16,9 +16,13 @@
         "ClusterId": "companyjobs",
         "Match": { "Path": "/jobs/{**catchall}" }
       },
-      "candidates-applications": {
-        "ClusterId": "candidates",
+      "applications-applications": {
+        "ClusterId": "applications",
         "Match": { "Path": "/applications/{**catchall}" }
+      },
+      "applications-interviews": {
+        "ClusterId": "applications",
+        "Match": { "Path": "/interviews/{**catchall}" }
       }
     },
     "Clusters": {
@@ -33,6 +37,13 @@
         "Destinations": {
           "candidates-primary": {
             "Address": "http://candidates.hireflow.svc.cluster.local"
+          }
+        }
+      },
+      "applications": {
+        "Destinations": {
+          "applications-primary": {
+            "Address": "http://applications.hireflow.svc.cluster.local"
           }
         }
       }
